feat: connection with ui-flex

+++ Domain/Dto/OrderDto.cs
@@ -4,6 +4,6 @@ public class OrderDto
 {
     public Guid Id { get; set; }
     public Dictionary<Guid, int> ProductQuantities { get; set; } = new Dictionary<Guid, int>();
-    public double TotalPrice { get; set; }
+    public int TotalPrice { get; set; }
     public DateTime OrderDate { get; set; }
 }
